AC: Split collect unit tests, remove debug_warn
